committing broken code for safe keeping, will fix it later

+++ src/dot2/dot2common.hpp
@@ -7,6 +7,8 @@
 #include <condition_variable>
 #include <thread>
 #include <sys/time.h>
+#include <sys/prctl.h>
+#include <unistd.h>
 
 
 /* common structures with c++ linkage */
@@ -305,7 +307,7 @@ extern "C"
 
 typedef struct 
 {
-    uint8_t length;
+    size_t length;
     uint8_t *octets;
 }OctetString;
